pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "book-club-be"
version = "0.1.0"
description = "Book Club Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "asyncpg>=0.30",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "redis>=5.2",
    "structlog>=24.4",
    "prometheus-client>=0.21",
    "sentry-sdk[fastapi]>=2.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.27",
    "factory-boy>=3.3",
    "ruff>=0.8",
    "mypy>=1.13",
    "bandit>=1.8",
    "pre-commit>=4.0",
    "aiosqlite>=0.20",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B", "A", "C4", "T20", "RUF"]
ignore = [
    "B008",   # FastAPI: Depends() calls in function defaults are intentional
    "N815",   # Pydantic: camelCase field aliases for API JSON compatibility
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN", "N803", "S107", "E501", "N806", "RUF059"]
"app/schemas/*" = ["N815"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["app/main.py", "app/routers/health.py", "tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
